hf_kyc_space/requirements.txt
flask
ultralytics
easyocr
deepface
tf-keras
opencv-python-headless
numpy
pillow
gunicorn
python-multipart